Черновой вариант common.js
