+++ src/main.js
@@ -1,4 +1,7 @@
 'use strict';
-require('./pictures.js');
-require('./resizer.js');
-require('./upload.js');
+
+define([
+  './pictures',
+  './resizer',
+  './upload',
+], function(pictures, resizer, upload) {});
